Create `utils/augmentations.py`
